Move WitnessCount and WitnessCountCtx to primitive
