+++ lib/wallet/src/Cardano/Wallet/Transaction.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE DuplicateRecordFields #-}
@@ -64,7 +63,6 @@ import Cardano.Address.Derivation
     )
 import Cardano.Address.Script
     ( KeyHash (..)
-    , KeyRole (..)
     , Script (..)
     , ScriptTemplate
     )
@@ -96,9 +94,6 @@ import Cardano.Wallet.Primitive.Types.AnyExplicitScripts
     )
 import Cardano.Wallet.Primitive.Types.Coin
     ( Coin (..)
-    )
-import Cardano.Wallet.Primitive.Types.Hash
-    ( Hash (..)
     )
 import Cardano.Wallet.Primitive.Types.PoolId
     ( PoolId
@@ -135,6 +130,14 @@ import Cardano.Wallet.Primitive.Types.Tx.TxOut
 import Cardano.Wallet.Primitive.Types.ValidityIntervalExplicit
     ( ValidityIntervalExplicit (..)
     )
+import Cardano.Wallet.Primitive.Types.WitnessCount
+    ( WitnessCount (..)
+    , WitnessCount (..)
+    , WitnessCountCtx (..)
+    , emptyWitnessCount
+    , emptyWitnessCount
+    , toKeyRole
+    )
 import Control.DeepSeq
     ( NFData (..)
     )
@@ -153,9 +156,6 @@ import Data.Monoid.Monus
 import Data.Text
     ( Text
     )
-import Data.Word
-    ( Word8
-    )
 import Fmt
     ( Buildable (..)
     , genericF
@@ -170,7 +170,6 @@ import Internal.Cardano.Write.Tx.SizeEstimation
 import qualified Cardano.Wallet.Primitive.Types.TokenMap as TokenMap
 import qualified Cardano.Wallet.Primitive.Types.Tx.TxOut as TxOut
 import qualified Data.Foldable as F
-import qualified Data.List as L
 import qualified Data.Map.Strict as Map
 
 data TransactionLayer (k :: Depth -> Type -> Type) ktype tx = TransactionLayer
@@ -334,51 +333,6 @@ data DelegationAction
 
 instance Buildable DelegationAction where
     build = genericF
-
-data WitnessCount = WitnessCount
-    { verificationKey :: Word8
-    , scripts :: [AnyExplicitScript]
-    , bootstrap :: Word8
-    }
-    deriving (Eq, Generic, Show)
-    deriving anyclass NFData
-
-emptyWitnessCount :: WitnessCount
-emptyWitnessCount = WitnessCount
-    { verificationKey = 0
-    , scripts = []
-    , bootstrap = 0
-    }
-
--- WitnessCount context is needed to differentiate verification keys present
--- in native scripts.
--- In shelley wallets they could be present due to only policy verification key.
--- In multisig wallet they could stem from payment, policy and delegation roles,
--- and as minting/burning and delegation support comes will be extended in additional
--- data attached in SharedWalletCtx to differentiate that.
--- WitnessCount is needed only during or after signing, in other phases it is not used.
-data WitnessCountCtx =
-      ShelleyWalletCtx KeyHash -- Policy
-    | SharedWalletCtx [KeyHash] -- Delegation key hashes of all cosigners
-    | AnyWitnessCountCtx
-    deriving (Eq, Generic, Show)
-    deriving anyclass NFData
-
-toKeyRole :: WitnessCountCtx -> Hash "VerificationKey" -> KeyRole
-toKeyRole witCtx (Hash key) = case witCtx of
-    ShelleyWalletCtx (KeyHash _ mypolicykey) ->
-        if key == mypolicykey then
-            Policy
-        else
-            Unknown
-    SharedWalletCtx stakingKeyHashes ->
-        let toStakeKey (KeyHash _ k) = k
-            isStakeKey = L.elem key (map toStakeKey stakingKeyHashes)
-        in if isStakeKey then
-               Delegation
-           else
-               Payment
-    AnyWitnessCountCtx -> Unknown
 
 data ErrMkTransaction
     =  ErrMkTransactionTxBodyError Text
